x/schema change location of some functions
